switch pages and modify post fetch from session storage
